refactor - package 이름 수정(step3 -> racingcar)
