+++ src/main/java/racingcar/CarController.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 import java.util.List;
 import java.util.Random;
