+++ src/main/java/racingcar/CarMain.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 public class CarMain {
     public static void main(String[] args) {
