+++ src/main/java/racingcar/InputView.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 import java.util.Scanner;
