+++ src/test/java/racingcar/CarControllerTest.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
