+++ src/test/java/racingcar/CarTest.java
@@ -1,4 +1,4 @@
-package step3;
+package racingcar;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
